Refactor from inheritance based to RowableContext
