+++ Source/Blazorise/Components/Row/Row.razor
@@ -2,8 +2,10 @@
 @inherits BaseRowableComponent
 <div id="@ElementId" class="@ClassNames" style="@StyleNames" @attributes="@Attributes">
     <CascadingValue Value="@this" IsFixed>
-        <CascadingValue Value="@GutterState">
-            @ChildContent
+        <CascadingValue Value="@RowableContext" IsFixed>
+            <CascadingValue Value="@GutterState">
+                @ChildContent
+            </CascadingValue>
         </CascadingValue>
     </CascadingValue>
 </div>
