+++ Source/Blazorise/Components/Typography/DescriptionList.razor.cs
@@ -10,15 +10,11 @@ namespace Blazorise;
 /// <summary>
 /// A description list is a list of items with a description or definition of each item.
 /// </summary>
-public partial class DescriptionList : BaseTypographyComponent, IRowableComponent
+public partial class DescriptionList : BaseTypographyComponent
 {
     #region Members
 
     private bool row;
-
-    private int spaceUsedByColumnables = 0;
-
-    private List<IColumnableComponent> columnables = new();
 
     #endregion
 
@@ -35,46 +31,14 @@ public partial class DescriptionList : BaseTypographyComponent, IRowableComponen
         base.BuildClasses( builder );
     }
 
-    /// <inheritdoc/>
-    public void NotifyColumnableInitialized( IColumnableComponent column )
-    {
-        if ( columnables is not null && !columnables.Contains( column ) )
-        {
-            columnables.Add( column );
-        }
-    }
-
-    /// <inheritdoc/>
-    public void NotifyColumnableRemoved( IColumnableComponent column )
-    {
-        if ( columnables is not null && columnables.Contains( column ) )
-        {
-            columnables.Remove( column );
-        }
-    }
-
-    /// <inheritdoc/>
-    public void ResetUsedSpace( IColumnableComponent column )
-    {
-        if ( column is not null && columnables.IndexOf( column ) <= 0 )
-            spaceUsedByColumnables = 0;
-    }
-
-    /// <inheritdoc/>
-    public void IncreaseUsedSpace( int space )
-    {
-        spaceUsedByColumnables += space;
-
-        if ( spaceUsedByColumnables > 12 )
-            spaceUsedByColumnables = 12;
-    }
-
     #endregion
 
     #region Properties
 
-    /// <inheritdoc/>
-    public int TotalUsedSpace => spaceUsedByColumnables;
+    /// <summary>
+    /// Gets the rowable context used to calculate used space by the columns.
+    /// </summary>
+    [Inject] protected IRowableContext RowableContext { get; set; }
 
     /// <summary>
     /// Specifies that description list will be arranged in a rows and columns.
